Video Maker Model/requirements.txt
lightning
gradio
torch
diffusers
transformers
sentence-transformers
opencv-python
ffmpeg-python
numpy
Pillow
